changed threshold_KEY to THRESHOLD_KEY
very cool. When the original code was written, I had minThreshold for most things, but THRESHOLD_KEY as it is now. Then John decided that it should be 'threshold' instead of 'minThreshold'. So I used the powers that be (ctrl+h) to make the change. But I made a mistake. I changed things to 'Threshold', when it needed to be 'threshold'. So I had to ctrl+h again (Threshold => threshold), and I forgot to be case sensitive, so THRESHOLD_KEY got converted into threshold_KEY. Now I have changed things back, and checkstyle will hopefully pass.